git Split e2e spec and add utility for setting text
